backend/analytics/requirements.txt
streamlit==1.39.0
pandas==2.2.3
httpx==0.27.2
python-dotenv==1.0.1
altair==5.4.1
